codex(web): extract preset source-policy store
